Completed dialogues now get saved in a xml file
A new XML file gets created, saving the IDs of every important dialogue completed.
The, each scene has variables for those dialogues, when a scene gets loaded it searches for the IDs of the respective dialogues of the scene in an array located at DialogueSystem.h to check if they have been completed.